refactor: useInitiator 拆分hook
